Add insert_module_instructions
